requirements.txt
numpy~=1.26.4
ruff
python-dotenv~=1.0.1